refactor test setup to use webpack/mocha/karma - borrowing setup from jss project.
