+++ browsers.json
@@ -1,0 +1,79 @@
+{
+  "BS_Safari": {
+    "base": "BrowserStack",
+    "os": "OS X",
+    "os_version": "Yosemite",
+    "browser": "safari",
+    "browser_version": "8.0"
+  },
+  "BS_Safari": {
+    "base": "BrowserStack",
+    "os": "OS X",
+    "os_version": "Mavericks",
+    "browser": "safari",
+    "browser_version": "7.0"
+  },
+  "BS_Chrome": {
+    "base": "BrowserStack",
+    "os": "OS X",
+    "os_version": "Yosemite",
+    "browser": "chrome",
+    "browser_version": "49.0"
+  },
+  "BS_Firefox": {
+    "base": "BrowserStack",
+    "os": "OS X",
+    "os_version": "Yosemite",
+    "browser": "firefox",
+    "browser_version": "45.0"
+  },
+  "BS_MobileSafari": {
+    "base": "BrowserStack",
+    "os": "ios",
+    "os_version": "8.0",
+    "browser": "iphone",
+    "real_mobile": false
+  },
+  "BS_MobileSafari": {
+    "base": "BrowserStack",
+    "os": "ios",
+    "os_version": "9.0",
+    "browser": "iphone",
+    "real_mobile": false
+  },
+  "BS_Chrome": {
+    "base": "BrowserStack",
+    "os": "Windows",
+    "os_version": "8.1",
+    "browser": "chrome",
+    "browser_version": "49.0"
+  },
+  "BS_Firefox": {
+    "base": "BrowserStack",
+    "os": "Windows",
+    "os_version": "8.1",
+    "browser": "firefox",
+    "browser_version": "45.0"
+  },
+  "BS_InternetExplorer9": {
+    "base": "BrowserStack",
+    "os": "Windows",
+    "os_version": "7",
+    "browser": "ie",
+    "browser_version": "9.0"
+  },
+  "BS_InternetExplorer10": {
+    "base": "BrowserStack",
+    "os": "Windows",
+    "os_version": "8",
+    "browser": "ie",
+    "browser_version": "10"
+  },
+  "BS_InternetExplorer11": {
+    "base": "BrowserStack",
+    "os": "Windows",
+    "os_version": "8.1",
+    "browser": "ie",
+    "browser_version": "11.0"
+  }
+}
