+++ karma.conf.js
@@ -1,0 +1,72 @@
+const assign = require('lodash.assign')
+const webpackConfig = require('./webpack.config')
+const browsers = require('./browsers')
+
+const isBench = process.env.BENCHMARK === 'true'
+const useCloud = process.env.USE_CLOUD === 'true'
+const browserStackUserName = process.env.BROWSER_STACK_USERNAME
+const browserStackAccessKey = process.env.BROWSER_STACK_ACCESS_KEY
+const isTravis = process.env.TRAVIS
+const travisBuildNumber = process.env.TRAVIS_BUILD_NUMBER
+const travisBuildId = process.env.TRAVIS_BUILD_ID
+const travisJobNumber = process.env.TRAVIS_JOB_NUMBER
+
+module.exports = (config) => {
+  config.set({
+    customLaunchers: browsers,
+    browsers: ['Chrome', 'Firefox', 'Safari'],
+    frameworks: ['mocha'],
+    files: [
+      'node_modules/es5-shim/es5-shim.js',
+      'node_modules/es5-shim/es5-sham.js',
+      'tests.webpack.js'
+    ],
+    preprocessors: {
+      'tests.webpack.js': ['webpack', 'sourcemap']
+    },
+    webpack: assign(webpackConfig, {
+      devtool: 'inline-source-map'
+    }),
+    webpackServer: {
+      noInfo: true
+    },
+    reporters: ['mocha', 'coverage'],
+    coverageReporter: {
+      dir: 'coverage',
+      file: 'coverage.json',
+      type: 'json'
+    }
+  })
+  if (isBench) {
+    assign(config, {
+      browsers: ['Chrome'],
+      frameworks: ['benchmark'],
+      files: ['benchmark/**/*.js'],
+      preprocessors: {'benchmark/**/*.js': ['webpack']},
+      reporters: ['benchmark']
+    })
+  }
+
+  if (useCloud) {
+    assign(config, {
+      browsers: Object.keys(browsers),
+      browserDisconnectTimeout: 10000,
+      browserDisconnectTolerance: 3,
+      browserNoActivityTimeout: 30000,
+      captureTimeout: 200000
+    })
+
+    config.browserStack = {
+      username: browserStackUserName,
+      accessKey: browserStackAccessKey,
+      pollingTimeout: 10000
+    }
+
+    if (isTravis) {
+      assign(config.browserStack, {
+        build: `TRAVIS #${travisBuildNumber} (${travisBuildId})`,
+        name: travisJobNumber
+      })
+    }
+  }
+}
