+++ package.json
@@ -18,40 +18,48 @@
     "style",
     "stylesheet"
   ],
-  "engines": {},
+  "engines": {
+    "node": ">=4.0.0 <6.0.0"
+  },
   "scripts": {
-    "all": "npm run lint && npm run build",
-    "test": "opener ./test/index.html",
+    "all": "npm run lint && npm run test && npm run build",
+    "clean": "rimraf ./lib/*",
     "cover": "istanbul cover test/index.js",
     "build": "npm run clean && npm run build:lib && npm run build:max && npm run build:min",
-    "clean": "rimraf ./lib/*",
     "build:lib": "babel src --out-dir lib",
     "build:max": "NODE_ENV=development webpack src/index.js dist/perdido.js",
     "build:min": "NODE_ENV=production webpack src/index.js dist/perdido.min.js",
-    "lint": "eslint ./src",
-    "prepublish": "npm run all && git push --tags"
+    "build:test": "webpack tests/index.js tmp/tests.js",
+    "lint": "eslint ./src ./tests",
+    "prepublish": "npm run all && git push --tags",
+    "test": "./node_modules/karma/bin/karma start --single-run",
+    "test:watch": "karma start",
+    "posttest": "[ -z \"$TRAVIS\" ] || codecov",
+    "codecov": "codecov",
+    "bench": "BENCHMARK=true npm run test"
   },
   "license": "MIT",
   "main": "./lib/index.js",
   "devDependencies": {
-    "babel": "^6.5.2",
     "babel-cli": "^6.9.0",
     "babel-core": "^6.9.1",
     "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
+    "babel-plugin-transform-inline-environment-variables": "^6.8.0",
     "babel-plugin-transform-runtime": "^6.9.0",
+    "babel-polyfill": "^6.9.1",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
+    "chai": "^3.5.0",
+    "codecov": "^1.0.1",
     "es5-shim": "4.5.8",
     "eslint": "^2.11.1",
     "eslint-config-airbnb": "^8.0.0",
     "eslint-config-jss": "^1.1.0",
     "eslint-plugin-import": "^1.8.1",
     "eslint-plugin-jsx-a11y": "^1.3.0",
-    "eslint-plugin-promise": "^1.3.2",
     "eslint-plugin-react": "^5.1.1",
-    "eslint-plugin-standard": "^1.3.2",
     "istanbul": "^0.4.3",
     "jss": "^3.11.1",
     "jss-camel-case": "^1.1.0",
@@ -59,8 +67,17 @@
     "jss-extend": "^1.2.1",
     "jss-nested": "^1.0.3",
     "jss-vendor-prefixer": "^2.0.0",
-    "opener": "^1.4.1",
-    "qunitjs": "2.0.0-rc1",
+    "karma": "^0.13.22",
+    "karma-browserstack-launcher": "^1.0.1",
+    "karma-chrome-launcher": "^1.0.1",
+    "karma-coverage": "^1.0.0",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-mocha": "^1.0.1",
+    "karma-mocha-reporter": "^2.0.3",
+    "karma-safari-launcher": "^1.0.0",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^1.7.0",
+    "mocha": "^2.5.3",
     "rimraf": "^2.5.2",
     "webpack": "^1.13.1"
   },
