--- test/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Perdido</title>
-  <link rel="stylesheet" href="../node_modules/qunitjs/qunit/qunit.css">
-</head>
-<body>
-  <div id="qunit"></div>
-  <div id="qunit-fixture"></div>
-  <script src="../node_modules/qunitjs/qunit/qunit.js"></script>
-  <script src="../node_modules/jss/dist/jss.js"></script>
-  <script src="../node_modules/jss-extend/dist/jss-extend.js"></script>
-  <script src="../node_modules/jss-nested/dist/jss-nested.js"></script>
-  <script src="../node_modules/jss-camel-case/dist/jss-camel-case.js"></script>
-  <script src="../node_modules/jss-default-unit/dist/jss-default-unit.js"></script>
-  <script src="../node_modules/jss-vendor-prefixer/dist/jss-vendor-prefixer.js"></script>
-  <script src="../dist/perdido.js"></script>
-  <script src="./index.js"></script>
-</body>
-</html>
