+++ tests.webpack.js
@@ -1,0 +1,2 @@
+const context = require.context('./test', true, /\.js$/)
+context.keys().forEach(context)
