+++ tests/failing.spec.js
@@ -1,0 +1,5 @@
+describe("A test suite", function() {
+  beforeEach(function() { });
+  afterEach(function() { });
+  it('should fail', function() { expect(true).to.be.false; });
+});
