+++ tests/index.html
@@ -1,0 +1,22 @@
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>Perdido Tests</title>
+  <link href="../node_modules/mocha/mocha.css" rel="stylesheet" />
+</head>
+<body>
+<div id="mocha"></div>
+<script src="../node_modules/es5-shim/es5-shim.js"></script>
+<script src="../node_modules/es5-shim/es5-sham.js"></script>
+
+<script src="../node_modules/expect.js/index.js"></script>
+<script src="../node_modules/mocha/mocha.js"></script>
+
+<script>mocha.setup('bdd')</script>
+<script src="../tmp/tests.js"></script>
+<script>
+  mocha.checkLeaks();
+  mocha.run();
+</script>
+</body>
+</html>
